allow using implementations only without running the generators + allow passing revised compilations
