+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/Generators/GodotPluginsInitializerGenerator.cs
@@ -1,0 +1,18 @@
+using Microsoft.CodeAnalysis;
+
+namespace Godot.SourceGenerators;
+
+[Generator]
+public class GodotPluginsInitializerGenerator : ISourceGenerator
+{
+    private readonly GodotPluginsInitializerGeneratorImplementation _implementation = new();
+
+    public void Initialize(GeneratorInitializationContext context)
+    {
+    }
+
+    public void Execute(Microsoft.CodeAnalysis.GeneratorExecutionContext context)
+    {
+        _implementation.Execute(new GeneratorExecutionContext(context));
+    }
+}
