+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/Generators/ScriptPropertiesGenerator.cs
@@ -1,0 +1,18 @@
+using Microsoft.CodeAnalysis;
+
+namespace Godot.SourceGenerators;
+
+[Generator]
+public class ScriptPropertiesGenerator : ISourceGenerator
+{
+    private readonly ScriptPropertiesGeneratorImplementation _implementation = new();
+
+    public void Initialize(GeneratorInitializationContext context)
+    {
+    }
+
+    public void Execute(Microsoft.CodeAnalysis.GeneratorExecutionContext context)
+    {
+        _implementation.Execute(new GeneratorExecutionContext(context));
+    }
+}
