+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/Generators/ScriptRegistrarGenerator.cs
@@ -1,0 +1,18 @@
+using Microsoft.CodeAnalysis;
+
+namespace Godot.SourceGenerators;
+
+// Placeholder. Once we switch to native extensions this will act as the registrar for all
+// user Godot classes in the assembly. Think of it as something similar to `register_types`.
+public class ScriptRegistrarGenerator : ISourceGenerator
+{
+    public void Initialize(GeneratorInitializationContext context)
+    {
+        throw new System.NotImplementedException();
+    }
+
+    public void Execute(Microsoft.CodeAnalysis.GeneratorExecutionContext context)
+    {
+        throw new System.NotImplementedException();
+    }
+}
