+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/ScriptPropertyDefValGeneratorImplementation.cs
@@ -8,14 +8,9 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Godot.SourceGenerators
 {
-    [Generator]
-    public class ScriptPropertyDefValGenerator : ISourceGenerator
+    public class ScriptPropertyDefValGeneratorImplementation
     {
-        public void Initialize(GeneratorInitializationContext context)
-        {
-        }
-
-        public void Execute(GeneratorExecutionContext context)
+        public void Execute(IGeneratorExecutionContext context)
         {
             if (context.IsGodotSourceGeneratorDisabled("ScriptPropertyDefVal"))
                 return;
@@ -58,7 +53,7 @@ namespace Godot.SourceGenerators
         }
 
         private static void VisitGodotScriptClass(
-            GeneratorExecutionContext context,
+            IGeneratorExecutionContext context,
             MarshalUtils.TypeCache typeCache,
             INamedTypeSymbol symbol
         )
