+++ modules/mono/editor/Godot.NET.Sdk/Godot.SourceGenerators/ScriptRegistrarGeneratorImplementation.cs
@@ -4,14 +4,14 @@ namespace Godot.SourceGenerators
 {
     // Placeholder. Once we switch to native extensions this will act as the registrar for all
     // user Godot classes in the assembly. Think of it as something similar to `register_types`.
-    public class ScriptRegistrarGenerator : ISourceGenerator
+    public class ScriptRegistrarGeneratorImplementation
     {
         public void Initialize(GeneratorInitializationContext context)
         {
             throw new System.NotImplementedException();
         }
 
-        public void Execute(GeneratorExecutionContext context)
+        public void Execute(IGeneratorExecutionContext context)
         {
             throw new System.NotImplementedException();
         }
